fix: Handle single image case in interactionCreate.ts
The code change in interactionCreate.ts now handles the case when there is only a single image to send. Previously, it assumed that the `images` variable would always be an array, but now it checks if `images` is a string. If it is, it creates an `AttachmentBuilder` with the image path and sends it to the user. Additionally, the code now properly unlinks the image path when it is done. This fix improves the functionality of sending single images in the Discord bot.